[hotfix][yarn] Move yarn Utils#calculateHeapSize to BootstrapTools
The method BootstrapTools#calculateHeapSize will be used by Yarn and Kubernetes to calculate the heapsize of jobmanager. The heapsize will be set to JVM -Xmx and -Xms.
